add identity server and enable authentication and authorization for auctionService

+++ src/AuctionService/Controllers/AuctionController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -48,12 +49,12 @@ public class AuctionController : ControllerBase
     return _mapper.Map<AuctionDto>(auction);
   }
 
+  [Authorize]
   [HttpPost]
   public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto) 
   {
     var auction = _mapper.Map<Auction>(auctionDto);
-    //TODO: add current user as seller
-    auction.Seller = "test";
+    auction.Seller = User.Identity.Name;
     _context.Auctions.Add(auction);
     var result = await _context.SaveChangesAsync() > 0;
 
@@ -65,7 +66,8 @@ public class AuctionController : ControllerBase
     
     return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, _mapper.Map<AuctionDto>(auction));
   }
-
+  
+  [Authorize]
   [HttpPut("{id}")]
   public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto auctionDto) 
   {
@@ -74,7 +76,7 @@ public class AuctionController : ControllerBase
 
     if (auction == null) return NotFound();
 
-    //TODO: check seller username
+    if (auction.Seller != User.Identity.Name) return Forbid();
 
     auction.Item.Make = auctionDto.Make ?? auction.Item.Make;
     auction.Item.Model = auctionDto.Model ?? auction.Item.Model;
@@ -91,6 +93,7 @@ public class AuctionController : ControllerBase
     return Ok();
   }
 
+  [Authorize]
   [HttpDelete("{id}")]
   public async Task<ActionResult> DeleteAuction(Guid id) 
   {
@@ -98,7 +101,7 @@ public class AuctionController : ControllerBase
 
     if (auction == null) return NotFound();
 
-    //TODO: check seller = username
+    if (auction.Seller != User.Identity.Name) return Forbid();
 
     _context.Auctions.Remove(auction);
